data-processing/requirements.txt
pyserial==3.5
easy_trilateration==0.1.4
ipykernel==6.23.0
pandas==2.0.1
plotly==5.14.1
nbformat==5.8.0